myapi/requirements.txt
flask
pymongo[srv]
flask-pymongo
flask-restful
mypy
python-dotenv